+++ lib/rdf/utils/regex.ex
@@ -1,20 +1,28 @@
-defmodule RDF.Util.Regex do
+defmodule RDF.Utils.Regex do
   @moduledoc """
-  Some of the regular expressions in the code base get executed in quote tight
-  loops, sometimes millions of times. The elixir code base does a check on every
-  run of a compiled regex, to check if the pcre version of the regex matches the
-  version that the local OTP version got compiled against. This check seems to
-  be unexpectedly costly (there is acall though to :erlang.system_info/1). It
-  seems worth it performance wise to circumvent elixir here, and call through to 
-  erlang 're' in this case.
+  Drop-in replacements for Elixir `Regex` functions.
+
+  Elixir does a check on every run of a compiled regex, to check if the PCRE
+  version of the regex matches the version that the local OTP version got
+  compiled against. This check seems to be unexpectedly costly (there is a call
+  though to `:erlang.system_info/1`). Since regular expression evaluation is
+  critical performance-wise in RDF.ex, especially during deserialization, all
+  regular expressions are evaluated through the functions in this module which
+  can be configured with the `optimize_regexes` key in the compile-time application
+  environment to circumvent Elixir and call through to Erlang `re` directly:
+
+      config :rdf,
+        optimize_regexes: true
+
+  By default this optimization is disabled and should be enabled only if your
+  application is running in a controlled environment.
   """
 
-  @doc """
-  Replaces the elixir regex run function. It only works with compiled regexes,
-  and always returns the matched binaries
-  """
-  @spec run(Regex.t(), binary()) :: nil | [binary]
   if Application.compile_env(:rdf, :optimize_regexes, false) do
+    @doc """
+    Drop-in replacement for `Regex.run/2` using Erlang `re` directly.
+    """
+    @spec run(Regex.t(), String.t()) :: nil | [binary]
     def run(%Regex{re_pattern: pattern}, string) do
       case :re.run(string, pattern, [{:capture, :all, :binary}]) do
         {:match, matches} -> matches
@@ -22,9 +30,16 @@ defmodule RDF.Util.Regex do
         :match -> []
       end
     end
-  else
-    def run(%Regex{} = re, string) do
-      Regex.run(re, string)
+
+    @doc """
+    Drop-in replacement for `Regex.match?/2` using Erlang `re` directly.
+    """
+    @spec match?(Regex.t(), String.t()) :: boolean
+    def match?(%Regex{re_pattern: pattern}, string) do
+      :re.run(string, pattern, [{:capture, :none}]) == :match
     end
+  else
+    defdelegate run(regex, string), to: Regex
+    defdelegate match?(regex, string), to: Regex
   end
 end